major/apt3090/group_project/requirements2.txt
pillow==10.3.0
pycryptodome==3.12.0
pycryptodomex==3.19.1